backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-dotenv>=1.0
sqlalchemy>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
argon2-cffi>=23.1
cryptography>=41.0
email-validator>=2.0
httpx>=0.25
python-multipart>=0.0.9
openai>=1.0
google-genai>=0.3.0
